+++ server/api_endpoints/api_endpoints.yml
@@ -719,11 +719,11 @@
   path: "/api/v1/fleet/assets"
   display_name: "List assets"
 - method: "GET"
-  path: "/api/v1/fleet/assets/:id"
+  path: "/api/v1/fleet/assets/:asset_uuid"
   display_name: "Get asset"
 - method: "POST"
   path: "/api/v1/fleet/assets"
   display_name: "Create asset"
 - method: "DELETE"
-  path: "/api/v1/fleet/assets/:id"
+  path: "/api/v1/fleet/assets/:asset_uuid"
   display_name: "Delete asset"
